+++ src/protocol/message.rs
@@ -6,7 +6,7 @@ use std::{
 };
 
 use super::frame::CloseFrame;
-use crate::error::{Error, Result};
+use crate::error::{CapacityError, Error, Result};
 
 mod string_collect {
     use utf8::DecodeError;
@@ -122,9 +122,10 @@ impl IncompleteMessage {
         let portion_size = tail.as_ref().len();
         // Be careful about integer overflows here.
         if my_size > max_size || portion_size > max_size - my_size {
-            return Err(Error::Capacity(
-                format!("Message too big: {} + {} > {}", my_size, portion_size, max_size).into(),
-            ));
+            return Err(Error::Capacity(CapacityError::MessageTooLong {
+                size: my_size + portion_size,
+                max_size,
+            }));
         }
 
         match self.collector {