+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.13.0
+
+- Add `CapacityError`, `UrlError`, and `ProtocolError` types to represent the different types of capacity, URL, and protocol errors respectively.
+- Modify variants `Error::Capacity`, `Error::Url`, and `Error::Protocol` to hold the above errors types instead of string error messages.
+
 # 0.12.0
 
 - Add facilities to allow clients to follow HTTP 3XX redirects.